feat(db): chatwoot_account_id + chatwoot_inbox_id to BIGINT (F2 MED-1)
PostgreSQL columns previously INTEGER would overflow when a Chatwoot
install auto-increments past 2^31. Chatwoot itself stores account and
inbox IDs as bigint, so the bridge must match.

Changes:
- migrations/0002_chatwoot_ids_bigint.sql: ALTER both columns to BIGINT
- Tenant and TenantInsert structs: int -> int64 for ChatwootAccountID
  and ChatwootInboxID
- CLI flags: IntVar -> Int64Var for --chatwoot-account / --chatwoot-inbox
- setupDB now applies all migrations in lexical order (was hardcoded to
  0001_init.sql) so test runs pick up future migrations automatically
- handlers_integration_test.go: replace JSON attachments expectation with
  multipart/form-data assertions for image + document inbound tests
  (these had been silently broken since the SCR-125 multipart refactor;
  they were testing the pre-multipart JSON pipeline that no longer exists)
- New TestInsertTenant_AcceptsLargeChatwootIDs exercises int64 with
  3_000_000_000 to prevent regression on the BIGINT migration

Tests: 102 unit + 126 integration pass.

+++ cmd/bridge/main.go
@@ -191,7 +191,7 @@ func cmdTenant(ctx context.Context, log zerolog.Logger, args []string) error {
 type tenantFlags struct {
 	slug, megaHost, megaInstance, megaToken string
 	cwURL, cwToken                          string
-	cwAccount, cwInbox                      int
+	cwAccount, cwInbox                      int64
 	skipReachCheck                          bool
 }
 
@@ -204,8 +204,8 @@ func parseTenantFlags(args []string) (tenantFlags, error) {
 	fs.StringVar(&f.megaToken, "megaapi-token", "", "megaAPI bearer token")
 	fs.StringVar(&f.cwURL, "chatwoot-url", "", "Chatwoot base URL")
 	fs.StringVar(&f.cwToken, "chatwoot-token", "", "Chatwoot api_access_token")
-	fs.IntVar(&f.cwAccount, "chatwoot-account", 0, "Chatwoot account id")
-	fs.IntVar(&f.cwInbox, "chatwoot-inbox", 0, "Chatwoot inbox id")
+	fs.Int64Var(&f.cwAccount, "chatwoot-account", 0, "Chatwoot account id")
+	fs.Int64Var(&f.cwInbox, "chatwoot-inbox", 0, "Chatwoot inbox id")
 	fs.BoolVar(&f.skipReachCheck, "skip-reach-check", false, "skip HEAD reachability check (test/dev)")
 	if err := fs.Parse(args); err != nil {
 		return f, err
